+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Regenerate exposed route data when new routes have been added after initial generation
 * Add `Conduit#setRouteParameters(parameters)` so that route parameters can be updated instead of totally replaced
 * Set a default `locale` option for new `Criteria` instances when a `Conduit` is available
+* Fix translatable field values not being queried correctly
 
 ## 1.2.8 (2022-11-13)